Move setupTerminal to run_common.go
